Renamed PyTorch Trainer files for consistency
