+++ juneberry/pytorch/privacy/trainer.py
@@ -27,7 +27,7 @@ import sys
 
 from opacus import PrivacyEngine
 
-from juneberry.pytorch.classifier_trainer import ClassifierTrainer
+from juneberry.pytorch.trainer import ClassifierTrainer
 
 logger = logging.getLogger(__name__)
